reels-script-generator/requirements.txt
python-dotenv==1.0.1
yt-dlp>=2024.1.0
